move InputsUsed to separate module
to exclude it from application to expression tools
